Implement Map Based Services. closes #6
Create the map package with the abstract class AbstractMapService that will use a HashMap to implement the persistence of data
sort of an in-memory DB with common CRUD operations.
Afterwards, create the services for Owner, Pet and Vet extending from the abstract class
but implementing the CrudService interface, which they match. Calling super for the impl.